Separated all proxy-server-for-integration-tests stuff into different util module
